Fix compilation problems in windows

+++ src/inet/applications/udpapp/UDPBasicP2P2B.cc
@@ -737,7 +737,7 @@ void UDPBasicP2P2B::handleMessage(cMessage *msg)
     {
         if (msg == &periodicTimer)
         {
-            emit(queueLengthSignal, timeQueue.size());
+            emit(queueLengthSignal, (unsigned long)timeQueue.size());
             scheduleAt(simTime()+par("periodicTimer"), &periodicTimer);
             return;
         }
